+++ src/main/java/net/starschema/clouddb/jdbc/BQScrollableResultSet.java
@@ -198,7 +198,7 @@ public class BQScrollableResultSet extends ScrollableResultset<Object> implement
                     return Boolean.parseBoolean(result);
                 }
                 if (Columntype.equals("INTEGER")) {
-                    return Integer.parseInt(result);
+                    return Long.parseLong(result);
                 }
                 if (Columntype.equals("TIMESTAMP")) {
                     long val = new BigDecimal(result).longValue() * 1000;
